Modified scripts and created output JSON
The restaurant pull script was modified to output a JSON file containing the results from the Google Places API pull. This file was then manually edited to contain the proper food categories and adjusted social good categories (UNREPRESENTATIVE OF ACTUAL RECTAURANTS; THIS IS A LIMITATION IN OUR DATA, AS THIS INFORMATION WAS UNGATHERABLE FOR EVERY RESTAURANT). Another script was then created to populate the MySQL database with data from the JSON file.